pivot from using an authorization grants, to using a consent endpoint handled inside of the oauth controller and using the authorize ednpoint as the step where the user is given the decision option, then the consent endpoint is where the user submits the decision and receives the redirect url.
