+++ spec/requests/oauth_controller_spec.rb
@@ -49,18 +49,31 @@ RSpec.describe Oauth::OauthController, type: :request do
     end
 
     context 'when request is valid' do
-      it 'returns a redirect_url with the correct parameters' do
+      it 'returns consent information' do
         make_request
         json = JSON.parse(response.body)
-        expect(json).to have_key('redirect_url')
 
-        uri = URI.parse(json['redirect_url'])
-        params = Rack::Utils.parse_query(uri.query)
+        expect(json).to have_key('consent_info')
+        consent_info = json['consent_info']
 
-        expect(uri.to_s).to start_with(valid_client.redirect_uri)
-        expect(params).to have_key('code')
-        # regex to match the pattern of the code(64 lowercase alphanumeric characters). does not compare the actual value itself.
-        expect(params['code']).to match(/\A[a-f0-9]{64}\z/)
+        expect(consent_info['client_name']).to eq(valid_client.client_name)
+        expect(consent_info['user_name']).to eq("#{valid_user.first_name} #{valid_user.last_name}")
+        expect(consent_info['requested_permissions']).to be_an(Array)
+        expect(consent_info['state']).to match(/\A[a-f0-9]{64}\z/)
+        expect(consent_info['decision_options']).to match({
+          'allow' => {
+            'body' => {
+              'state' => consent_info['state'],
+              'decision' => 'allow'
+            }
+          },
+          'deny' => {
+            'body' => {
+              'state' => consent_info['state'],
+              'decision' => 'deny'
+            }
+          }
+        })
       end
     end
 
@@ -136,6 +149,119 @@ RSpec.describe Oauth::OauthController, type: :request do
         let(:request_params) { { code_challenge_method: 'plain' } }
         let(:error_message) { 'Invalid code_challenge_method' }
         it_behaves_like 'an invalid request'
+      end
+    end
+  end
+
+  describe 'POST /oauth/consent' do
+    let(:valid_state_token) {
+      state_data = {
+        client_id: valid_client.client_id,
+        redirect_uri: valid_client.redirect_uri,
+        user_id: valid_user.id,
+        code_challenge: 'test_challenge_123',
+        code_challenge_method: 'S256',
+        created_at: Time.current.iso8601
+      }
+
+      state_token = SecureRandom.hex(32)
+      Redis.current.setex("oauth_state:#{state_token}", 600, state_data.to_json)
+      state_token
+    }
+
+    def make_consent_request(state: valid_state_token, decision: 'allow')
+      post "/oauth/consent",
+           params: { state: state, decision: decision }.to_json,
+           headers: { 'Content-Type' => 'application/json' }
+    end
+
+    shared_examples 'an invalid consent request' do
+      it 'returns a :bad_request status' do
+        make_consent_request(**request_params)
+        expect(response).to have_http_status(:bad_request)
+      end
+
+      it 'returns a JSON response with the error message' do
+        make_consent_request(**request_params)
+        expect(JSON.parse(response.body)['error']).to eq(error_message)
+      end
+    end
+
+    context 'when user allows access' do
+      it 'returns a redirect_url with authorization code' do
+        make_consent_request(decision: 'allow')
+
+        expect(response).to have_http_status(:ok)
+        json = JSON.parse(response.body)
+        expect(json).to have_key('redirect_url')
+
+        uri = URI.parse(json['redirect_url'])
+        params = Rack::Utils.parse_query(uri.query)
+
+        expect(uri.to_s).to start_with(valid_client.redirect_uri)
+        expect(params).to have_key('code')
+        expect(params['code']).to match(/\A[a-f0-9]{64}\z/)
+      end
+
+      it 'deletes the state token from Redis' do
+        state_key = "oauth_state:#{valid_state_token}"
+        expect(Redis.current.exists(state_key)).to eq(1)
+
+        make_consent_request(decision: 'allow')
+
+        expect(Redis.current.exists(state_key)).to eq(0)
+      end
+    end
+
+    context 'when user denies access' do
+      it 'returns a redirect_url with access_denied error' do
+        make_consent_request(decision: 'deny')
+
+        expect(response).to have_http_status(:ok)
+        json = JSON.parse(response.body)
+        expect(json).to have_key('redirect_url')
+
+        uri = URI.parse(json['redirect_url'])
+        params = Rack::Utils.parse_query(uri.query)
+
+        expect(uri.to_s).to start_with(valid_client.redirect_uri)
+        expect(params['error']).to eq('access_denied')
+        expect(params['error_description']).to eq('User denied authorization')
+      end
+
+      it 'deletes the state token from Redis' do
+        state_key = "oauth_state:#{valid_state_token}"
+        expect(Redis.current.exists(state_key)).to eq(1)
+
+        make_consent_request(decision: 'deny')
+
+        expect(Redis.current.exists(state_key)).to eq(0)
+      end
+    end
+
+    context 'when request is invalid' do
+      context 'when state is missing' do
+        let(:request_params) { { state: nil } }
+        let(:error_message) { 'Missing state parameter' }
+        it_behaves_like 'an invalid consent request'
+      end
+
+      context 'when state is invalid' do
+        let(:request_params) { { state: 'invalid_state' } }
+        let(:error_message) { 'Invalid or expired state token' }
+        it_behaves_like 'an invalid consent request'
+      end
+
+      context 'when decision is missing' do
+        let(:request_params) { { decision: nil } }
+        let(:error_message) { 'Missing decision parameter' }
+        it_behaves_like 'an invalid consent request'
+      end
+
+      context 'when decision is invalid' do
+        let(:request_params) { { decision: 'maybe' } }
+        let(:error_message) { 'Invalid decision' }
+        it_behaves_like 'an invalid consent request'
       end
     end
   end
